bug fixes, selector feature preparation
fixed memory leak when using operator=(ValueItem&&) or operator=(const ValueItem&)
updated bytecode to be more cleary and implicit
now BuildCall is eager build instead lazy function build
reduced warns in run_time_compiler.cpp

+++ run_time/standard_lib.cpp
@@ -670,7 +670,7 @@ ValueItem math_frrt_impl(ValueItem& val) {
 
 	switch (val.meta.vtype) {
 	case VType::flo: {
-		float f = t;
+		float f = (float)t;
 		return ValueItem((void*)&f, ValueMeta(VType::flo, false, true));
 	}
 	default:
